restructure handler, service and store
