+++ api/resources/handler/resource/handler_test.go
@@ -1,10 +1,11 @@
-package handler
+package resource
 
 import (
 	"bytes"
 	"context"
 	"errors"
 	"fmt"
+	"github.com/zopdev/zopdev/api/resources/service/resource"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -16,7 +17,6 @@ import (
 	gofrHttp "gofr.dev/pkg/gofr/http"
 
 	"github.com/zopdev/zopdev/api/resources/models"
-	"github.com/zopdev/zopdev/api/resources/service"
 )
 
 var errMock = errors.New("mock error")
@@ -108,7 +108,7 @@ func TestHandler_ChangeState(t *testing.T) {
 	ctx := &gofr.Context{
 		Context: context.Background(),
 	}
-	resDetails := service.ResourceDetails{CloudAccID: 123, Name: "sql-instance-1", Type: "sql", State: service.START}
+	resDetails := resource.ResourceDetails{CloudAccID: 123, Name: "sql-instance-1", Type: "sql", State: resource.START}
 	h := New(mockSvc)
 
 	testCases := []struct {
