+++ api/resources/service/resource/cron.go
@@ -1,4 +1,4 @@
-package service
+package resource
 
 import (
 	"sync"
