+++ api/resources/service/resource/cron_test.go
@@ -1,4 +1,4 @@
-package service
+package resource
 
 import (
 	"context"
