+++ api/resources/service/resource/interface.go
@@ -1,4 +1,4 @@
-package service
+package resource
 
 import (
 	"context"
