+++ api/resources/service/resource/models.go
@@ -1,4 +1,4 @@
-package service
+package resource
 
 type (
 	CloudProvider string
