+++ api/resources/service/resource/service_test.go
@@ -1,4 +1,4 @@
-package service
+package resource
 
 import (
 	"context"
