+++ api/resources/service/resource/statechange.go
@@ -1,4 +1,4 @@
-package service
+package resource
 
 import (
 	"strings"
