+++ api/resources/store/resource/store_test.go
@@ -1,4 +1,4 @@
-package store
+package resource
 
 import (
 	"context"
